models/ecg_jepa/requirements.txt
einops==0.8.0
h5py==3.11.0
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.1
pickle5==0.0.11
PyYAML==6.0.1
scikit_learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
timm==1.0.8
torchmetrics==1.4.0.post0
tqdm==4.66.4
wfdb==4.1.2
